chore(license): relicense to Apache-2.0 with LLVM Exception
This is the standard LLVM license and thus the license for clangd itself.
I've had an informal request to use this license for clangd-adjacent
stuff where we can, to make cross-contributions easier.

It's a slightly more liberal version of the Apache 2 license.
The exception section was designed with compilers in mind, and is
included for uniformity rather than because it's important per se.